requirements.txt

requests
Pillow
python-dotenv
gradio
